+++ backend/DTOs/UserGameStatDto.cs
@@ -6,7 +6,7 @@ namespace DailyChallenges.DTOs
         public string Name { get; set; } = string.Empty;
         public string? Url { get; set; }
         public int Plays { get; set; }
-        public int? HighestScore { get; set; }
+        public double? HighestScore { get; set; }
         public int? BestSubmissionId { get; set; }
         public DateTime? LastPlayed { get; set; }
         public string? RankingMode { get; set; }
